fases e moedas
adicionei mais fases, adicionei as moedas, arrumei o script pra aceitar moedas com valor decimal, adicionei as fases na build

+++ Assets/Scripts/MoneyItem.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public class MoneyItem : MonoBehaviour
 {
-    public int value; // Define o valor da nota/moeda
+    public float value; // Define o valor da nota/moeda
 }
